==> 1-data_collection_and_cleaning.ipynb <==
# Dataset

import pandas as pd

dataset_path = "datasets/creditcard_original.csv"
df = pd.read_csv(dataset_path)

df.head()

Dataset Source: https://www.kaggle.com/datasets/mlg-ulb/creditcardfraud

Features:
* Time: time of the transaction
* Amount: amount of the transaction
* V1 to V28: confidential variables for the transaction

Target: Class
* 1: Fraudulent Transactions 
* 0: Non Fraudulent Transactions 

# Data Cleaning

### Handling Missing Values

df.isnull().sum()

### Remove Duplicates

df.duplicated().sum()

df = df.drop_duplicates()

df.duplicated().sum()

## Exporting Cleaned Dataset

df.to_csv("datasets/creditcard_cleaned.csv", index=False)
