==> 3-data_preprocessing.ipynb <==
# Data Preprocessing

import pandas as pd

df = pd.read_csv("datasets/creditcard_cleaned.csv")

## Feature Selection

V10, V12, V14 and V17 have strong negative correlations with the "Class" label.

V4 and V11 have strong positive correlations with the "Class" label.

Therefore we will choose theses features for training our models.

features = ["V4", "V10", "V11", "V12", "V14", "V17"]

## Handling Outliers

print(f"DataFrame shape before dropping outliers: {df.shape}")
rows_before = df.shape[0]

for feature in features:
    Q1 = df[feature].quantile(0.25)
    Q3 = df[feature].quantile(0.75)
    IQR = Q3 - Q1

    # Define the lower and upper bounds for outliers
    lower_bound = Q1 - 1.5 * IQR
    upper_bound = Q3 + 1.5 * IQR

    # Filter out the outliers
    df = df[(df[feature] >= lower_bound) & (df[feature] <= upper_bound)]

print(f"DataFrame shape after dropping outliers: {df.shape}")
print(f"Rows Dropped: {rows_before - df.shape[0]}")

## Selecting Features and Target

x = df[features]
y = df["Class"]

## Handling Unbalanced Data

Use SMOTE Oversampling for the minority

from imblearn.over_sampling import SMOTE

smote = SMOTE(sampling_strategy="minority", random_state=42, k_neighbors=5)
x_res, y_res = smote.fit_resample(x, y)

### Original vs Resampled

print(f"Original class distribution: {y.value_counts()}")
print(f"Resampled class distribution: {y_res.value_counts()}")

## Exporting Balanced Dataset

df_res = pd.concat([x_res, y_res], axis=1)

df_res.to_csv("datasets/creditcard_balanced.csv", index=False)
