==> 4-machine_learning_models.ipynb <==
# Machine Learning

## Preparation

### Loading the Dataset

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

df = pd.read_csv("datasets/creditcard_balanced.csv")

df.head()

### Selecting Features and Target

x = df.drop("Class", axis=1)
y = df["Class"]

### Data Split to train and test

from sklearn.model_selection import train_test_split

x_train, x_test, y_train, y_test = train_test_split(
    x, y, test_size=0.3, random_state=42
)

### Classifiction Metrics

from sklearn.metrics import (
    classification_report,
    accuracy_score,
    # recall_score,
    # precision_score,
    # f1_score,
    confusion_matrix,
    roc_auc_score,
)

# from sklearn.metrics._plot.confusion_matrix import (
#     confusion_matrix,
# )


def print_metrics(y_true, y_pred):
    print("Classification Report: \n", classification_report(y_true, y_pred))
    print("Model Accuracy: ", accuracy_score(y_true, y_pred))
    print("ROC AUC: ", roc_auc_score(y_true, y_pred))
    print("Confusion matrix: \n", confusion_matrix(y_true, y_pred))

### Grid Search for Hyperparameter Tuning

from sklearn.model_selection import GridSearchCV


def grid_search(model, params, scoring="accuracy", cv=5):
    grid = GridSearchCV(model, params, scoring=scoring, cv=cv)
    grid.fit(x_train, y_train)
    print("Best params: ", grid.best_params_)
    return grid.best_estimator_

## Logistic Regression

### Model Training

from sklearn.linear_model import LogisticRegression

model_lr = LogisticRegression(random_state=42)

model_lr.fit(x_train, y_train)

### Model Prediction & Metrics

y_pred_lr = model_lr.predict(x_test)

print_metrics(y_test, y_pred_lr)

## KNN

### Model Training & Hyperparameter Tuning

from sklearn.neighbors import KNeighborsClassifier

params = {"n_neighbors": [3, 5, 7, 9], "p": [1, 2]}

model_knn = grid_search(KNeighborsClassifier(), params)

### Model Prediction & Metrics

y_pred_knn = model_knn.predict(x_test)

print_metrics(y_test, y_pred_knn)

## Decision Tree Classifier

### Model Training & Hyperparameter Tuning

from sklearn.tree import DecisionTreeClassifier

params = {"criterion": ["entropy", "gini"], "max_depth": range(3, 11)}

model_dtc = grid_search(DecisionTreeClassifier(random_state=42), params)

### Model Prediction & Metrics

y_pred_dtc = model_dtc.predict(x_test)

print_metrics(y_test, y_pred_dtc)

## Random Forest

### Model Training & Hyperparameter Tuning

from sklearn.ensemble import RandomForestClassifier

params = {"n_estimators": [370, 375, 380]}

model_rfc = grid_search(RandomForestClassifier(random_state=42), params)

### Model Prediction & Metrics

y_pred_rfc = model_rfc.predict(x_test)

print_metrics(y_test, y_pred_rfc)

# Exporting Models

import pickle
import joblib

### Logistic Regression

with open("models/logistic_regression_model.pkl", "wb") as file:
    pickle.dump(model_lr, file)

joblib.dump(model_lr, "models/logistic_regression_model.joblib")

### KNN

with open("models/knn_model.pkl", "wb") as file:
    pickle.dump(model_knn, file)

joblib.dump(model_knn, "models/knn_model.joblib")

### Decision Tree Classifier

with open("models/dtc_model.pkl", "wb") as file:
    pickle.dump(model_dtc, file)

joblib.dump(model_dtc, "models/dtc_model.joblib")

### Random Forest

with open("models/rfc_model.pkl", "wb") as file:
    pickle.dump(model_rfc, file)

joblib.dump(model_rfc, "models/rfc_model.joblib")
